Fixes: Bux Fixes and Code Refactoring
