+++ app/src/main/java/io/budge/trackme/TrackerActivity.kt
@@ -1,9 +1,6 @@
 package io.budge.trackme
 
-import android.animation.ObjectAnimator
-import android.animation.TypeEvaluator
 import android.os.Bundle
-import android.util.Property
 import android.widget.Toast
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
@@ -16,10 +13,11 @@ import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.maps.model.Marker
 import com.google.android.gms.maps.model.MarkerOptions
 import io.budge.trackme.data.User
+import io.budge.trackme.utils.Constants.IP_ADDRESS
+import io.budge.trackme.utils.Constants.PORT
+import io.budge.trackme.utils.animateMarkerToPosition
 import io.budge.trackme.utils.getAddress
 import kotlinx.android.synthetic.main.activity_tracker.*
-import kotlin.math.abs
-import kotlin.math.sign
 
 
 class TrackerActivity :
@@ -34,8 +32,6 @@ class TrackerActivity :
     private var markersMap = mutableMapOf<Int, Marker>()
 
     private lateinit var toast: Toast
-
-    private var openOnResume = false
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -65,22 +61,26 @@ class TrackerActivity :
 
         viewModel.locationUpdate.observe(this, {
             it?.let {
-                val marker = markersMap[it.id]
-                val user = marker?.tag as User
-                user.location = it
-                user.address = getAddress(it.lat, it.lng, this)
-                marker.tag = user
-                markersMap[it.id] = marker
-                if (marker.isInfoWindowShown) {
-                    marker.hideInfoWindow()
-                }
-                val finalPosition = LatLng(it.lat, it.lng)
-                animateMarker(marker, finalPosition)
-                val cameraUpdate = CameraUpdateFactory.newLatLngZoom(finalPosition, 15f)
-                mMap.animateCamera(cameraUpdate)
+                updateUsersLocations(it)
             }
         })
 
+    }
+
+    private fun updateUsersLocations(it: User.UserLocation) {
+        val marker = markersMap[it.id]
+        val user = marker?.tag as User
+        user.address = LatLng(it.lat, it.lng).getAddress(this)
+        user.location = it
+        marker.tag = user
+        markersMap[it.id] = marker
+        if (marker.isInfoWindowShown) {
+            marker.hideInfoWindow()
+        }
+        val finalPosition = LatLng(it.lat, it.lng)
+        marker.animateMarkerToPosition(finalPosition)
+        val cameraUpdate = CameraUpdateFactory.newLatLngZoom(finalPosition, 15f)
+        mMap.animateCamera(cameraUpdate)
     }
 
     private fun showToast(message: String){
@@ -88,46 +88,13 @@ class TrackerActivity :
         toast.show()
     }
 
-    private fun interpolate(fraction: Float, a: LatLng, b: LatLng): LatLng {
-        val lat = (b.latitude - a.latitude) * fraction + a.latitude
-        var lngDelta = b.longitude - a.longitude
-
-        if (abs(lngDelta) > 180) {
-            lngDelta -= sign(lngDelta) * 360
-        }
-        val lng = lngDelta * fraction + a.longitude
-        return LatLng(lat, lng)
-    }
-
-
-    private fun animateMarker(
-        marker: Marker?,
-        finalPosition: LatLng?,
-    ) {
-        val typeEvaluator: TypeEvaluator<LatLng> =
-            TypeEvaluator { fraction, startValue, endValue ->
-                interpolate(
-                    fraction,
-                    startValue,
-                    endValue
-                )
-            }
-        val property: Property<Marker, LatLng> =
-            Property.of(Marker::class.java, LatLng::class.java, "position")
-        val animator = ObjectAnimator.ofObject(marker, property, typeEvaluator, finalPosition)
-        animator.duration = 500
-        animator.start()
-    }
-
 
     override fun onMapReady(googleMap: GoogleMap) {
         mMap = googleMap
         mMap.setInfoWindowAdapter(CustomInfoWindowAdapter(this))
         viewModel.startConnection(
-            "ios-test.printful.lv",
-            6111,
-            "jerryb.adeleye@gmail.com"
-        )
+            IP_ADDRESS,
+            PORT)
     }
 
     private fun addMarkers(users: MutableList<User>) {
@@ -149,23 +116,23 @@ class TrackerActivity :
 
     override fun onPause() {
         super.onPause()
-        openOnResume = true
         for (marker in markersMap.values) marker.remove()
         viewModel.closeConnection()
+        finish()
     }
 
-    override fun onResume() {
-        super.onResume()
-        if (openOnResume) {
-            val cameraUpdate = CameraUpdateFactory.newLatLngZoom(LatLng(0.0,0.0), 1f)
-            mMap.animateCamera(cameraUpdate)
-            viewModel.startConnection(
-                "ios-test.printful.lv",
-                6111,
-                "jerryb.adeleye@gmail.com"
-            )
-        }
-    }
+//    override fun onResume() {
+//        super.onResume()
+//        if (openOnResume) {
+//            val cameraUpdate = CameraUpdateFactory.newLatLngZoom(LatLng(0.0,0.0), 1f)
+//            mMap.animateCamera(cameraUpdate)
+//            viewModel.startConnection(
+//                "ios-test.printful.lv",
+//                6111,
+//                "jerryb.adeleye@gmail.com"
+//            )
+//        }
+//    }
 
     override fun onMarkerClick(p0: Marker?): Boolean {
         return false
