+++ app/src/main/java/io/budge/trackme/TrackerViewModel.kt
@@ -8,10 +8,12 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import io.budge.trackme.data.Result
 import io.budge.trackme.data.User
+import io.budge.trackme.utils.Constants.INVALID_MESSAGE_FORMAT
+import io.budge.trackme.utils.Constants.UPDATE_PREFIX
+import io.budge.trackme.utils.Constants.USER_LIST_PREFIX
 import io.budge.trackme.utils.MessageProcessor
-import kotlinx.coroutines.Dispatchers
+import io.budge.trackme.utils.SocketManager
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 
 
 class TrackerViewModel(application: Application) : AndroidViewModel(application) {
@@ -40,6 +42,7 @@ class TrackerViewModel(application: Application) : AndroidViewModel(application)
             when (result) {
                 is Result.Success -> {
                     e("Success", result.data)
+                    _isLoading.postValue(false)
                     process(result)
                 }
                 is Result.Error -> {
@@ -55,30 +58,30 @@ class TrackerViewModel(application: Application) : AndroidViewModel(application)
     private fun process(result: Result.Success<Any>) {
         e("process", result.data)
         viewModelScope.launch {
-            val isUserList = result.data.startsWith("USERLIST")
-            val isUpdate = result.data.startsWith("UPDATE")
+            val isUserList = result.data.startsWith(USER_LIST_PREFIX)
+            val isUpdate = result.data.startsWith(UPDATE_PREFIX)
             try {
                 if (isUserList) {
-                    val users = messageProcessor.processUserList(result.data)
+                    val usersString = result.data.removePrefix(USER_LIST_PREFIX).trim().removeSuffix(";")
+                    val users = messageProcessor.processUserList(usersString)
                     _userList.value = users
                 } else if (isUpdate) {
-                    val locationUpdate = messageProcessor.processLocationUpdate(result.data)
+                    val updateString = result.data.removePrefix(UPDATE_PREFIX).trim()
+                    val locationUpdate = messageProcessor.processLocationUpdate(updateString)
                     _locationUpdate.value = locationUpdate
                 }
             } catch (e: NumberFormatException) {
-                _errorMessage.postValue("Invalid response format")
+                _errorMessage.postValue(INVALID_MESSAGE_FORMAT)
                 _isLoading.postValue(false)
             }
         }
     }
 
-    fun startConnection(ipAddress: String, port: Int, emailAddress: String) {
+    fun startConnection(ipAddress: String, port: Int) {
         _isLoading.postValue(true)
         e("startConnection", "start")
         viewModelScope.launch {
-            withContext(Dispatchers.IO) {
-                socketManager.openConnection(ipAddress, port, emailAddress)
-            }
+            socketManager.openConnection(ipAddress, port)
         }
     }
