+++ app/src/main/java/io/budge/trackme/utils/Constants.kt
@@ -1,0 +1,11 @@
+package io.budge.trackme.utils
+
+object Constants {
+
+    const val IP_ADDRESS = "ios-test.printful.lv"
+    const val PORT = 6111
+    const val UPDATE_PREFIX = "UPDATE"
+    const val USER_LIST_PREFIX = "USERLIST"
+    const val INVALID_MESSAGE_FORMAT = "Invalid response format"
+    const val AUTHORIZE_COMMAND = "AUTHORIZE jerryb.adeleye@gmail.com"
+}
