+++ app/src/main/java/io/budge/trackme/utils/MessageProcessor.kt
@@ -1,18 +1,12 @@
 package io.budge.trackme.utils
 
 import android.content.Context
-import android.graphics.Bitmap
-import android.graphics.BitmapFactory
-import android.location.Address
-import android.location.Geocoder
 import android.util.Log.e
+import com.google.android.gms.maps.model.LatLng
 import io.budge.trackme.data.User
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.withContext
-import java.io.InputStream
-import java.net.URL
-import java.util.*
 
 class MessageProcessor(private val context: Context) {
 
@@ -20,13 +14,12 @@ class MessageProcessor(private val context: Context) {
     suspend fun processUserList(message: String): MutableList<User> {
         val users = mutableListOf<User>()
         withContext(Dispatchers.IO) {
-            val usersString = message.removePrefix("USERLIST").trim().removeSuffix(";")
-            val usersStringList = usersString.split(";")
+            val usersStringList = message.split(";")
             for (userString in usersStringList) {
                 val userObject = userString.split(",")
                 val imageUrl = userObject[2].trim()
 
-                val imageBitmap = async {getBitmap(imageUrl)}
+                val imageBitmap = async {imageUrl.getBitmapFromStringUrl()}
                 val lat = userObject[3].trim().toDouble()
                 val lng = userObject[4].trim().toDouble()
                 val location = User.UserLocation(
@@ -38,7 +31,7 @@ class MessageProcessor(private val context: Context) {
                     userObject[0].trim().toInt(),
                     userObject[1].trim(),
                     imageUrl,
-                    getAddress(lat, lng, context),
+                    LatLng(lat, lng).getAddress(context),
                     imageBitmap.await(),
                     location
                 )
@@ -52,8 +45,7 @@ class MessageProcessor(private val context: Context) {
 
     @Throws(NumberFormatException::class)
     fun processLocationUpdate(message: String): User.UserLocation {
-        val updateString = message.removePrefix("UPDATE").trim()
-        val updateObject = updateString.split(",")
+        val updateObject = message.split(",")
         val id = updateObject[0].trim().toInt()
         return User.UserLocation(
             id,
@@ -61,25 +53,4 @@ class MessageProcessor(private val context: Context) {
             updateObject[2].trim().toDouble()
         )
     }
-
-    private fun getBitmap(imageUrl: String): Bitmap? {
-        val url = URL(imageUrl)
-        val inputStream: InputStream = url.openStream()
-        return BitmapFactory.decodeStream(inputStream)
-    }
-
-
-}
-
-fun getAddress(lat: Double, lng: Double, context: Context): String {
-    val addresses: List<Address>
-    val geoCoder = Geocoder(context, Locale.getDefault())
-
-    addresses = geoCoder.getFromLocation(
-        lat,
-        lng,
-        1)
-
-    val address = addresses[0]
-    return address.getAddressLine(0)
 }
